+++ app/src/main/java/com/example/taskschedulerproject/NotesList.java
@@ -11,8 +11,8 @@ public class NotesList extends ItemList {
     public void addItem(String title,
                         String description,
                         Date date,
-                        short priority,
-                        short duration,
+                        int priority,
+                        int duration,
                         Date deadline) {
         ArrayList<Item> allNotes = getItems();
         allNotes.add(new NoteItem(title, description, date));
